requirements.txt
dash==2.15.0
dash_bootstrap_components==1.5.0
ipython==8.12.3
numpy==1.26.4
pandas==2.2.1
plotly==5.19.0
pymongo==4.6.2
python-dotenv==1.0.1
gunicorn==21.2.0
